parse to double

+++ src/serverRequest.js
@@ -12,7 +12,7 @@ document.addEventListener("DOMContentLoaded", function () {
         request.addEventListener('readystatechange', function () {
             console.log(parseInt(request.responseText.replace(/[^\d]/g, '')));
             document.querySelector("#img").classList.add('opacity');
-            document.querySelector("#result").innerHTML = `${parseInt(request.responseText.replace(/[^\d]/g, ''))} % unique`;
+            document.querySelector("#result").innerHTML = `${parseDouble(request.responseText.replace(/[^\d]/g, ''))} % unique`;
         });
 
         request.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded');
